# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screenshot capture API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Viewport type
 */
export type ScreenshotInputViewport = typeof ScreenshotInputViewport[keyof typeof ScreenshotInputViewport];


export const ScreenshotInputViewport = {
  desktop: 'desktop',
  mobile: 'mobile',
} as const;

export interface ScreenshotInput {
  /** Full URL to screenshot */
  url: string;
  /** Viewport type */
  viewport?: ScreenshotInputViewport;
  /** Capture full page scroll height */
  fullPage?: boolean;
}

export type BulkScreenshotInputViewport = typeof BulkScreenshotInputViewport[keyof typeof BulkScreenshotInputViewport];


export const BulkScreenshotInputViewport = {
  desktop: 'desktop',
  mobile: 'mobile',
} as const;

export interface BulkScreenshotInput {
  /** Base URL of the website */
  baseUrl: string;
  /** List of paths to screenshot (e.g. ["/", "/about", "/contact"]) */
  paths: string[];
  viewport?: BulkScreenshotInputViewport;
  fullPage?: boolean;
}

export interface DiscoverInput {
  /** URL to discover pages from */
  url: string;
}

export interface DiscoveredPage {
  url: string;
  path: string;
  label: string;
}

export interface DiscoveredPages {
  baseUrl: string;
  pages: DiscoveredPage[];
}

export type ScreenshotResultViewport = typeof ScreenshotResultViewport[keyof typeof ScreenshotResultViewport];


export const ScreenshotResultViewport = {
  desktop: 'desktop',
  mobile: 'mobile',
} as const;

export type ScreenshotResultStatus = typeof ScreenshotResultStatus[keyof typeof ScreenshotResultStatus];


export const ScreenshotResultStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ScreenshotResult {
  id: string;
  url: string;
  viewport: ScreenshotResultViewport;
  fullPage: boolean;
  status: ScreenshotResultStatus;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export type BulkJobStatus = typeof BulkJobStatus[keyof typeof BulkJobStatus];


export const BulkJobStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface BulkJob {
  jobId: string;
  status: BulkJobStatus;
  total: number;
  completed: number;
  failed: number;
  /** @nullable */
  zipUrl?: string | null;
  screenshots: ScreenshotResult[];
  createdAt: string;
}