# lib/api-zod/src/generated/types/bulkJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screenshot capture API
 * OpenAPI spec version: 0.1.0
 */
import type { BulkJobStatus } from './bulkJobStatus';
import type { ScreenshotResult } from './screenshotResult';

export interface BulkJob {
  jobId: string;
  status: BulkJobStatus;
  total: number;
  completed: number;
  failed: number;
  /** @nullable */
  zipUrl?: string | null;
  screenshots: ScreenshotResult[];
  createdAt: string;
}