# lib/api-zod/src/generated/types/bulkScreenshotInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screenshot capture API
 * OpenAPI spec version: 0.1.0
 */
import type { BulkScreenshotInputViewport } from './bulkScreenshotInputViewport';

export interface BulkScreenshotInput {
  /** Base URL of the website */
  baseUrl: string;
  /** List of paths to screenshot (e.g. ["/", "/about", "/contact"]) */
  paths: string[];
  viewport?: BulkScreenshotInputViewport;
  fullPage?: boolean;
}