# lib/api-zod/src/generated/types/discoverInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screenshot capture API
 * OpenAPI spec version: 0.1.0
 */

export interface DiscoverInput {
  /** URL to discover pages from */
  url: string;
}