# lib/api-zod/src/generated/types/discoveredPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screenshot capture API
 * OpenAPI spec version: 0.1.0
 */

export interface DiscoveredPage {
  url: string;
  path: string;
  label: string;
}