# lib/api-zod/src/generated/types/discoveredPages.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screenshot capture API
 * OpenAPI spec version: 0.1.0
 */
import type { DiscoveredPage } from './discoveredPage';

export interface DiscoveredPages {
  baseUrl: string;
  pages: DiscoveredPage[];
}