# lib/api-zod/src/generated/types/screenshotInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screenshot capture API
 * OpenAPI spec version: 0.1.0
 */
import type { ScreenshotInputViewport } from './screenshotInputViewport';

export interface ScreenshotInput {
  /** Full URL to screenshot */
  url: string;
  /** Viewport type */
  viewport?: ScreenshotInputViewport;
  /** Capture full page scroll height */
  fullPage?: boolean;
}