# lib/api-zod/src/generated/types/screenshotResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screenshot capture API
 * OpenAPI spec version: 0.1.0
 */
import type { ScreenshotResultStatus } from './screenshotResultStatus';
import type { ScreenshotResultViewport } from './screenshotResultViewport';

export interface ScreenshotResult {
  id: string;
  url: string;
  viewport: ScreenshotResultViewport;
  fullPage: boolean;
  status: ScreenshotResultStatus;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}